Gradle version with android studio Work as Standalone Application
