+++ EM_Decipherment/src/main/java/com/example/decipher/Bag.java
@@ -1,0 +1,188 @@
+package com.example.decipher;
+/*
+ Compilation:  javac Bag.java
+ Execution:    java Bag < input.txt
+ Dependencies: StdIn.java StdOut.java
+
+ A generic bag or multiset, implemented using a singly-linked list.
+
+ % more tobe.txt
+ to be or not to - be - - that - - - is
+
+ % java Bag < tobe.txt
+ size of bag = 14
+ is
+ -
+ -
+ -
+ that
+ -
+ -
+ be
+ -
+ to
+ not
+ or
+ be
+ to
+
+ */
+
+
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+
+/**
+ * The <tt>Bag</tt> class represents a bag (or multiset) of
+ * generic items. It supports insertion and iterating over the
+ * items in arbitrary order.
+ * <p>
+ * This implementation uses a singly-linked list with a static nested class Node.
+ * See {@link LinkedBag} for the version from the
+ * textbook that uses a non-static nested class.
+ * The <em>add</em>, <em>isEmpty</em>, and <em>size</em> operations
+ * take constant time. Iteration takes time proportional to the number of items.
+ * <p>
+ * For additional documentation, see <a href="http://algs4.cs.princeton.edu/13stacks">Section 1.3</a> of
+ * <i>Algorithms, 4th Edition</i> by Robert Sedgewick and Kevin Wayne.
+ *
+ * @param <Item> the generic type of an item in this bag
+ * @author Robert Sedgewick
+ * @author Kevin Wayne
+ */
+public class Bag<Item> implements Iterable<Item> {
+	private Node<Item> first;    // beginning of bag
+	private int N;               // number of elements in bag
+
+	/**
+	 * helper linked list class
+	 *
+	 * @param <Item>
+	 */
+	private static class Node<Item> {
+		private Item item;
+		private Node<Item> next;
+	}
+
+	/**
+	 * Initializes an empty bag.
+	 */
+	public Bag() {
+		first = null;
+		N = 0;
+	}
+
+	/**
+	 * Returns true if this bag is empty.
+	 *
+	 * @return <tt>true</tt> if this bag is empty;
+	 * <tt>false</tt> otherwise
+	 */
+	public boolean isEmpty() {
+		return first == null;
+	}
+
+	/**
+	 * Returns the number of items in this bag.
+	 *
+	 * @return the number of items in this bag
+	 */
+	public int size() {
+		return N;
+	}
+
+	/**
+	 * Adds the item to this bag.
+	 *
+	 * @param item the item to add to this bag
+	 */
+	public void add(Item item) {
+		Node<Item> oldfirst = first;
+		first = new Node<>();
+		first.item = item;
+		first.next = oldfirst;
+		N++;
+	}
+
+
+	/**
+	 * Returns an iterator that iterates over the items in this bag in arbitrary order.
+	 *
+	 * @return an iterator that iterates over the items in this bag in arbitrary order
+	 */
+	public Iterator<Item> iterator() {
+		return new ListIterator<>(first);
+	}
+
+	//
+
+	/**
+	 * an iterator, doesn't implement remove() since it's optional
+	 *
+	 * @param <Item>
+	 */
+	private class ListIterator<Item> implements Iterator<Item> {
+		private Node<Item> current;
+
+		public ListIterator(Node<Item> first) {
+			current = first;
+		}
+
+		public boolean hasNext() {
+			return current != null;
+		}
+
+		public void remove() {
+			throw new UnsupportedOperationException();
+		}
+
+		public Item next() {
+			if (!hasNext()) throw new NoSuchElementException();
+			Item item = current.item;
+			current = current.next;
+			return item;
+		}
+	}
+
+	/**
+	 * Unit tests the <tt>Bag</tt> data type.
+	 */
+	public static void main(String[] args) {
+		Bag<String> bag = new Bag<>();
+		while (!StdIn.isEmpty()) {
+			String item = StdIn.readString();
+			bag.add(item);
+		}
+
+		StdOut.println("size of bag = " + bag.size());
+		for (String s : bag) {
+			StdOut.println(s);
+		}
+	}
+
+
+}
+
+/*
+ Copyright 2002-2015, Robert Sedgewick and Kevin Wayne.
+
+ This file is part of algs4.jar, which accompanies the textbook
+
+ Algorithms, 4th edition by Robert Sedgewick and Kevin Wayne,
+ Addison-Wesley Professional, 2011, ISBN 0-321-57351-X.
+ http://algs4.cs.princeton.edu
+
+
+ algs4.jar is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ algs4.jar is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with algs4.jar.  If not, see http://www.gnu.org/licenses.
+ */
