+++ EM_Decipherment/src/test/java/com/example/decipher/LinearProbingHashSTTest.java
@@ -1,0 +1,12 @@
+package com.example.decipher;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class LinearProbingHashSTTest {
+
+	@Test
+	public void main() {
+	}
+}
