+++ phonetic_corp.txt
